requests: fix decoding of single empty parameter
This fixes an inconsistency where
- "AAAA PARM" has no parameters,
- "AAAA PARM " also has no parameters, but
- "AAAA PARM ;" has two empty parameters.

While there probably is no practical use, this is an actual bug in
decoding as we either should never have allowed empty parameters
in the protocol (which we do and may need in some special cases).
It does not make sense to be able to only be able to end up with
2 or more empty parameters but not a single one although syntax is
clear on what was meant.